% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korrelation.R
\name{correlation_coefficients}
\alias{correlation_coefficients}
\alias{cor_pearson}
\alias{cor_spearman}
\alias{cor_kendall}
\title{Correlation coefficients}
\usage{
cor_pearson(x, y)

cor_spearman(x, y)

cor_kendall(x, y)
}
\arguments{
\item{x, y}{Vectors.}
}
\description{
Wrapper to \link[stats:cor]{stats::cor}.
}
