% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streuungsparameter.R
\name{empirical_deviation}
\alias{empirical_deviation}
\alias{var_n}
\alias{sd_n}
\alias{cov_n}
\title{Empirical deviations}
\usage{
var_n(x)

sd_n(x)

cov_n(x, y)
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\description{
Empirical deviations
}
