% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haeufigkeiten.R
\name{grouped_data}
\alias{grouped_data}
\title{Summarise grouped data}
\usage{
grouped_data(breaks, x, x_type = c("x", "h_j", "f_j"), n)
}
\arguments{
\item{breaks}{A numeric vector with length \emph{n + 1} where \emph{n} is the number of
groups. Group \emph{i} is the right-open interval [\code{breaks[i]}, \code{breaks[i + 1]}).
Group \emph{n} is right-closed.}

\item{x}{A numeric vector. See section \code{x_type}.}

\item{x_type}{See section \code{x_type}.}

\item{n}{Sample size (only used if \code{x_type = "f_j"})}
}
\description{
Summarise grouped data
}
\section{x_type}{


Based on \code{x_type} the argument \code{x} is treated as follows:
\itemize{
\item \code{x} : Values to be grouped.
\item \code{h_j} : Absolute frequencies of each group. Then
\code{length(x) == length(breaks) - 1}.
\item \code{f_j} : Relative frequencies of each group. Then
\code{length(x) == length(breaks) - 1}. In this case the sample size \code{n} has
to be specified.
}
}

