% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konzentration.R
\name{herfindahl_index}
\alias{herfindahl_index}
\alias{modified_herfindahl_index}
\title{Herfindahl index}
\usage{
herfindahl_index(x)

modified_herfindahl_index(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Herfindahl index

Modified Herfindahl index
}
