% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konzentration.R
\name{lorenz_concentration}
\alias{lorenz_concentration}
\title{Lorenz concentration curve}
\usage{
lorenz_concentration(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Table containing data constituting the Lorenz concentration curve.
}
