% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mittelwerte.R
\name{means}
\alias{means}
\alias{harmonic_mean}
\alias{geometric_mean}
\title{Further means}
\usage{
harmonic_mean(x)

geometric_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Further means
}
