% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konzentration.R
\name{modified_gini_coefficient}
\alias{modified_gini_coefficient}
\title{Modified Gini coefficient}
\usage{
modified_gini_coefficient(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Modified Gini coefficient
}
