% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konzentration.R
\name{modified_herfindahl_index}
\alias{modified_herfindahl_index}
\title{Modified Herfindahl index}
\usage{
modified_herfindahl_index(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Modified Herfindahl index
}
