% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schiefe.R
\name{quantile_skewness}
\alias{quantile_skewness}
\title{Quantile skewness}
\usage{
quantile_skewness(x, p = 0.25)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{Probability (less than 0.5)}
}
\description{
Quantile skewness
}
