% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haeufigkeiten.R
\name{sorted_to_vector}
\alias{sorted_to_vector}
\title{Transform contingency table of counts to vector}
\usage{
sorted_to_vector(x, count)
}
\arguments{
\item{x}{A numeric vector with unique values.}

\item{count}{A numeric vector with same length as \code{x}.}
}
\description{
Transform contingency table of counts to vector
}
